requirements.txt
# streamlit
# qrcode
# pillow
# Flask
# pandas
# openpyxl


streamlit
qrcode
pillow
requests
pandas
numpy